Add files via upload
In this project I have merged all the 12 months files into a single file and then finds the product which has the maximum sales with the specific months. I also find the  which product sold most often and which products are sold together as-well for the business benefits in future.